requirements.txt
xgboost
jieba
gensim
pyyaml
sklearn
pandas
tensorflow>=1.10
